+++ src/modules/reply.js
@@ -10,9 +10,8 @@ module.exports = bot => {
     if (! thread) return;
 
     const text = args.join(' ').trim();
-    if (msg.attachments.length) await attachments.saveAttachmentsInMessage(msg);
-    await thread.replyToUser(msg.member, text, msg.attachments, false);
-    msg.delete();
+    const replied = await thread.replyToUser(msg.member, text, msg.attachments, false);
+    if (replied) msg.delete();
   });
 
   bot.registerCommandAlias('r', 'reply');
@@ -22,9 +21,8 @@ module.exports = bot => {
     if (! thread) return;
 
     const text = args.join(' ').trim();
-    if (msg.attachments.length) await attachments.saveAttachmentsInMessage(msg);
-    await thread.replyToUser(msg.member, text, msg.attachments, true);
-    msg.delete();
+    const replied = await thread.replyToUser(msg.member, text, msg.attachments, true);
+    if (replied) msg.delete();
   });
 
   bot.registerCommandAlias('ar', 'anonreply');
